* commit '32624f29f9006da16e83d9a23745453f5f2416dd':
  BluetoothMidiService: Add support for sending SysEx messages that span multiple Bluetooth packets

+++ core/java/com/android/internal/midi/MidiConstants.java
@@ -87,13 +87,13 @@ public final class MidiConstants {
     }
 
     // Returns true if this command can be used for running status
-    public static boolean allowRunningStatus(int command) {
+    public static boolean allowRunningStatus(byte command) {
         // only Channel Voice and Channel Mode commands can use running status
         return (command >= STATUS_NOTE_OFF && command < STATUS_SYSTEM_EXCLUSIVE);
     }
 
     // Returns true if this command cancels running status
-    public static boolean cancelsRunningStatus(int command) {
+    public static boolean cancelsRunningStatus(byte command) {
         // System Common messages cancel running status
         return (command >= STATUS_SYSTEM_EXCLUSIVE && command <= STATUS_END_SYSEX);
     }
